backend/requirements.txt
fastapi
uvicorn
python-multipart
motor
pymongo
pydantic[email]
email-validator
python-dotenv
passlib[bcrypt]
pyjwt
aiofiles
openai
boto3
botocore
celery[redis]
redis
pdfplumber
python-docx
faiss-cpu
numpy
pytesseract
Pillow
pdf2image
mammoth
six>=1.16.0
pdfminer.six==20250506
reportlab
openpyxl
xlrd
authlib
httpx
sentence-transformers
torch
